requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.3.2
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
